+++ Kernel/Modules/AdminOTRSBusiness.pm
@@ -127,25 +127,47 @@ sub Run {
         );
     }
     elsif ( $Self->{Subaction} eq 'InstallAction' ) {
-        if ( $OTRSBusinessObject->OTRSBusinessInstall() ) {
+        my %Response = $OTRSBusinessObject->OTRSBusinessInstall();
+        if ( $Response{Success} ) {
             return $LayoutObject->Redirect(
                 OP => "Action=AdminOTRSBusiness;NotificationCode=InstallOk"
             );
         }
 
+        my $Parameters = '';
+        if ( $Response{RequiredFrameworkMinimum} ) {
+            $Parameters .= ";RequiredFrameworkMinimum=$Response{RequiredFrameworkMinimum}";
+        }
+        if ( $Response{RequiredFrameworkMaximum} ) {
+            $Parameters .= ";RequiredFrameworkMaximum=$Response{RequiredFrameworkMaximum}";
+        }
+        if ( $Response{ShowBlock} ) {
+            $Parameters .= ";ShowBlock=$Response{ShowBlock}";
+        }
         return $LayoutObject->Redirect(
-            OP => "Action=AdminOTRSBusiness;NotificationCode=InstallError"
+            OP => "Action=AdminOTRSBusiness;NotificationCode=InstallError" . $Parameters,
         );
     }
     elsif ( $Self->{Subaction} eq 'UpdateAction' ) {
-        if ( $OTRSBusinessObject->OTRSBusinessUpdate() ) {
+        my %Response = $OTRSBusinessObject->OTRSBusinessUpdate();
+        if ( $Response{Success} ) {
             return $LayoutObject->Redirect(
                 OP => "Action=AdminOTRSBusiness;NotificationCode=UpdateOk"
             );
         }
 
+        my $Parameters = '';
+        if ( $Response{RequiredFrameworkMinimum} ) {
+            $Parameters .= ";RequiredFrameworkMinimum=$Response{RequiredFrameworkMinimum}";
+        }
+        if ( $Response{RequiredFrameworkMaximum} ) {
+            $Parameters .= ";RequiredFrameworkMaximum=$Response{RequiredFrameworkMaximum}";
+        }
+        if ( $Response{ShowBlock} ) {
+            $Parameters .= ";ShowBlock=$Response{ShowBlock}";
+        }
         return $LayoutObject->Redirect(
-            OP => "Action=AdminOTRSBusiness;NotificationCode=UpdateError"
+            OP => "Action=AdminOTRSBusiness;NotificationCode=UpdateError" . $Parameters,
         );
     }
 
@@ -205,8 +227,26 @@ sub NotInstalledScreen {
         );
     }
     else {
+        my $ParamObject = $Kernel::OM->Get('Kernel::System::Web::Request');
+        my %GetParam;
+        $GetParam{ShowBlock}              = $ParamObject->GetParam( Param => 'ShowBlock' )                || '';
+        $GetParam{RequiredMinimumVersion} = $ParamObject->GetParam( Param => 'RequiredFrameworkMinimum' ) || '';
+        $GetParam{RequiredMaximumVersion} = $ParamObject->GetParam( Param => 'RequiredFrameworkMaximum' ) || '';
+
+        my $NotificationCode = $ParamObject->GetParam( Param => 'NotificationCode' ) || '';
+        if ( $NotificationCode eq 'InstallError' ) {
+            $GetParam{Type} = 'InstallIncompatible';
+        }
+
+        if ( $GetParam{Type} ) {
+            $LayoutObject->Block(
+                Name => 'Actions',
+            );
+        }
+
         $LayoutObject->Block(
             Name => 'Install',
+            Data => \%GetParam,
         );
     }
 
@@ -262,8 +302,27 @@ sub InstalledScreen {
             }
         }
         elsif ( $OTRSBusinessObject->OTRSBusinessIsUpdateable() ) {
+            my $ParamObject = $Kernel::OM->Get('Kernel::System::Web::Request');
+            my %GetParam;
+            $GetParam{ShowBlock}              = $ParamObject->GetParam( Param => 'ShowBlock' )                || '';
+            $GetParam{RequiredMinimumVersion} = $ParamObject->GetParam( Param => 'RequiredFrameworkMinimum' ) || '';
+            $GetParam{RequiredMaximumVersion} = $ParamObject->GetParam( Param => 'RequiredFrameworkMaximum' ) || '';
+
+            my $NotificationCode = $ParamObject->GetParam( Param => 'NotificationCode' ) || '';
+
+            if ( $NotificationCode eq 'UpdateError' ) {
+                $GetParam{Type} = 'UpgradeIncompatible';
+            }
+
+            if ( $GetParam{Type} ) {
+                $LayoutObject->Block(
+                    Name => 'Actions',
+                );
+            }
+
             $LayoutObject->Block(
                 Name => 'NeedsUpdate',
+                Data => \%GetParam,
             );
         }
         else {