+++ Kernel/System/OTRSBusiness.pm
@@ -686,15 +686,43 @@ downloads and installs OTRSBusiness.
 sub OTRSBusinessInstall {
     my ( $Self, %Param ) = @_;
 
+    my %Response = (
+        Success => 0,
+    );
+
     my $PackageString = $Self->_OTRSBusinessFileGet();
-    return if !$PackageString;
+    return %Response if !$PackageString;
+
+    my $PackageObject = $Kernel::OM->Get('Kernel::System::Package');
+
+    # Parse package structure.
+    my %Structure = $PackageObject->PackageParse(
+        String    => $PackageString,
+        FromCloud => 1,
+    );
+
+    my $FrameworkCheck = $PackageObject->_CheckFramework(
+        Framework  => $Structure{Framework},
+        NoLog      => 1,
+        ResultType => 'HASH',
+    );
+
+    # Check result type of _CheckFramework() for compatibility reasons.
+    if (
+        ref $FrameworkCheck eq 'HASH'
+        && !$FrameworkCheck->{Success}
+        )
+    {
+        $FrameworkCheck->{ShowBlock} = 'IncompatibleInfo';
+        return %{$FrameworkCheck};
+    }
 
     my $Install = $Kernel::OM->Get('Kernel::System::Package')->PackageInstall(
         String    => $PackageString,
         FromCloud => 1,
     );
 
-    return $Install if !$Install;
+    return %Response if !$Install;
 
     # now that we know that OTRSBusiness has been installed,
     # we can just preset the cache instead of just swiping it.
@@ -705,7 +733,9 @@ sub OTRSBusinessInstall {
         Value => 1,
     );
 
-    return $Install;
+    $Response{Success} = 1;
+
+    return %Response;
 }
 
 =item OTRSBusinessReinstall()
@@ -741,13 +771,46 @@ downloads and updates OTRSBusiness.
 sub OTRSBusinessUpdate {
     my ( $Self, %Param ) = @_;
 
+    my %Response = (
+        Success => 0,
+    );
+
     my $PackageString = $Self->_OTRSBusinessFileGet();
     return if !$PackageString;
 
-    return $Kernel::OM->Get('Kernel::System::Package')->PackageUpgrade(
+    my $PackageObject = $Kernel::OM->Get('Kernel::System::Package');
+
+    # Parse package structure.
+    my %Structure = $PackageObject->PackageParse(
         String    => $PackageString,
         FromCloud => 1,
     );
+
+    my $FrameworkCheck = $PackageObject->_CheckFramework(
+        Framework  => $Structure{Framework},
+        NoLog      => 1,
+        ResultType => 'HASH',
+    );
+
+    # Check result type of _CheckFramework() for compatibility reasons.
+    if (
+        ref $FrameworkCheck eq 'HASH'
+        && !$FrameworkCheck->{Success}
+        )
+    {
+        $FrameworkCheck->{ShowBlock} = 'IncompatibleInfo';
+        return %{$FrameworkCheck};
+    }
+
+    my $Upgrade = $Kernel::OM->Get('Kernel::System::Package')->PackageUpgrade(
+        String    => $PackageString,
+        FromCloud => 1,
+    );
+
+    return %Response if !$Upgrade;
+
+    $Response{Success} = 1;
+    return %Response;
 }
 
 =item OTRSBusinessUninstall()